requirements.txt
# asgiref==3.3.1
boto3==1.17.22
# botocore==1.20.22
# certifi==2020.12.5
# chardet==4.0.0
Django==3.1.7
# django-cachalot==2.3.3
# django-debug-toolbar==3.2
# django-model-utils==4.1.1
# django-storages==1.11.1
# fontawesome-free==5.15.2
# idna==2.10
# jmespath==0.10.0
# mysqlclient @ file:///C:/Users/Adam/Desktop/AIM/aim_writersdigest/mysqlclient-1.4.6-cp36-cp36m-win_amd64.whl
# ordered-set==4.0.2
Pillow==8.1.2
# python-dateutil==2.8.1
# pytz==2021.1
# requests==2.25.1
# s3transfer==0.3.4
# six==1.15.0
# sqlparse==0.4.1
# urllib3==1.26.3


boto3
Django==3.1.7
django-model-utils
django-storages
fontawesome-free
jmespath
ordered-set
Pillow
python-dateutil
pytz
s3transfer
six
sqlparse
urllib3
mysqlclient==2.0.1
django-debug-toolbar
django-cachalot